requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
anthropic==0.18.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0